+++ astroquery/irsa_dust/tests/test_irsa_dust_remote.py
@@ -1,15 +1,12 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-import imp
 import os
 import astropy.units as u
 import pytest
-import requests
 
 from astropy.tests.helper import remote_data
 from astropy.table import Table
 from ... import irsa_dust
 
-imp.reload(requests)
 
 M31_XML = "dustm31.xml"
 M81_XML = "dustm81.xml"