+++ astroquery/utils/commons.py
@@ -364,6 +364,9 @@ class FileContainer(object):
         """
         filedata = self.get_string()
 
+        if len(filedata) == 0:
+            raise TypeError("The file retrieved was empty.")
+
         self._fits = fits.HDUList.fromstring(filedata)
 
         return self._fits