+++ astroquery/vo_conesearch/conesearch.py
@@ -1,13 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""
-Support VO Simple Cone Search capabilities.
-
-.. note:: This maintains a similar API as ``astropy.vo.client``.
-
-"""
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-import six
+"""Support VO Simple Cone Search capabilities."""
 
 # STDLIB
 import warnings
@@ -23,8 +15,9 @@ from astropy.utils.exceptions import AstropyUserWarning
 # LOCAL
 from . import vos_catalog
 from .vo_async import AsyncBase
-from .core import ConeSearchClass, _validate_sr
-from .exceptions import ConeSearchError, VOSError
+from .core import ConeSearch, _validate_sr
+from .exceptions import ConeSearchError
+from ..exceptions import NoResultsWarning
 from ..utils.timer import timefunc, RunTimePredictor
 
 # Import configurable items declared in __init__.py
@@ -60,7 +53,7 @@ class AsyncConeSearch(AsyncBase):
     >>> c = coord.ICRS(6.0223 * u.degree, -72.0814 * u.degree)
     >>> async_search = conesearch.AsyncConeSearch(
     ...     c, 0.5 * u.degree,
-    ...     catalog_db='The PMM USNO-A1.0 Catalogue (Monet 1997) 1')
+    ...     catalog_db='Guide Star Catalog 2.3 Cone Search 1')
 
     Check search status:
 
@@ -78,20 +71,32 @@ class AsyncConeSearch(AsyncBase):
     If no ``timeout`` keyword given, it waits until completion:
 
     >>> async_result = async_search.get(timeout=30)
-    >>> cone_arr = async_result.array.data
-    >>> cone_arr.size
-    36184
+    >>> len(async_result)
+    74271
 
     """
     def __init__(self, *args, **kwargs):
         super(AsyncConeSearch, self).__init__(conesearch, *args, **kwargs)
 
 
-def conesearch(center, radius, verb=1, catalog_db=None, pedantic=None,
-               verbose=True, cache=True, timeout=None, query_all=False):
+def conesearch(center, radius, *, verb=1, catalog_db=None,
+               verbose=True, cache=True, query_all=False,
+               return_astropy_table=True, use_names_over_ids=False):
     """
     Perform Cone Search and returns the result of the
     first successful query.
+
+    .. note::
+
+        Use ``astroquery.vo_conesearch.conf.pedantic`` to control
+        pedantry. When `True`, will raise an error when the result
+        violates the spec, otherwise issue a warning. Warnings may
+        be controlled using :py:mod:`warnings` module.
+
+    .. note::
+
+        Use ``astroquery.vo_conesearch.conf.timeout`` to control
+        timeout limit in seconds for each service being queried.
 
     Parameters
     ----------
@@ -157,13 +162,6 @@ def conesearch(center, radius, verb=1, catalog_db=None, pedantic=None,
             - Any of the above 3 options combined in a list, in which case
               they are tried in order.
 
-    pedantic : bool or `None`
-        When `True`, raise an error when the result violates the spec,
-        otherwise issue a warning. Warnings may be controlled using
-        :py:mod:`warnings` module. When not provided, uses the
-        configuration setting ``astroquery.vo_conesearch.conf.pedantic``,
-        which defaults to `False`.
-
     verbose : bool
         Verbose output.
 
@@ -172,31 +170,32 @@ def conesearch(center, radius, verb=1, catalog_db=None, pedantic=None,
         websites referenced by the database still needs internet
         connection.
 
-    timeout : number or `None`
-        Timeout limit in seconds for each service being queries.
-        If `None`, use default.
-
     query_all : bool
         This is used by :func:`search_all`.
 
+    return_astropy_table : bool
+        Returned ``obj`` will be `astropy.table.Table` rather
+        than `astropy.io.votable.tree.Table`.
+
+    use_names_over_ids : bool
+        When `True` use the ``name`` attributes of columns as the names
+        of columns in the `~astropy.table.Table` instance.  Since names
+        are not guaranteed to be unique, this may cause some columns
+        to be renamed by appending numbers to the end.  Otherwise
+        (default), use the ID attributes as the column names.
+
     Returns
     -------
-    obj : `astropy.io.votable.tree.Table`
+    obj : `astropy.table.Table` or `astropy.io.votable.tree.Table`
         First table from first successful VO service request.
+        See ``return_astropy_table`` parameter for the kind of table returned.
 
     Raises
     ------
     ConeSearchError
         When invalid inputs are passed into Cone Search.
 
-    VOSError
-        If VO service request fails.
-
     """
-    # Not using default ConeSearch instance because the attributes are
-    # tweaked to match user inputs to this function.
-    cs = ConeSearchClass()
-
     n_timed_out = 0
     service_type = conf.conesearch_dbname
     catalogs = vos_catalog._get_catalogs(
@@ -206,13 +205,8 @@ def conesearch(center, radius, verb=1, catalog_db=None, pedantic=None,
     else:
         result = None
 
-    if pedantic is not None:
-        cs.PEDANTIC = pedantic
-    if timeout is not None:
-        cs.TIMEOUT = timeout
-
     for name, catalog in catalogs:
-        if isinstance(catalog, six.string_types):
+        if isinstance(catalog, str):
             if catalog.startswith('http'):
                 url = catalog
             else:
@@ -223,32 +217,31 @@ def conesearch(center, radius, verb=1, catalog_db=None, pedantic=None,
         else:
             url = catalog['url']
 
-        cs.URL = url
-
         if verbose:  # pragma: no cover
             color_print('Trying {0}'.format(url), 'green')
 
         try:
-            r = cs.query_region(center, radius, verb=verb, cache=cache,
-                                verbose=verbose)
-
+            r = ConeSearch.query_region(
+                center, radius, verb=verb, cache=cache, verbose=verbose,
+                service_url=url, return_astropy_table=return_astropy_table,
+                use_names_over_ids=use_names_over_ids)
         except Exception as e:
             err_msg = str(e)
             vo_warn(W25, (url, err_msg))
             if not query_all and 'ConnectTimeoutError' in err_msg:
                 n_timed_out += 1
         else:
-            if query_all:
-                result[r.url] = r
-            else:
-                result = r
-                break
+            if r is not None:
+                if query_all:
+                    result[r.url] = r
+                else:
+                    result = r
+                    break
 
-    if result is None:
-        err_msg = 'None of the available catalogs returned valid results.'
-        if n_timed_out > 0:
-            err_msg += ' ({0} URL(s) timed out.)'.format(n_timed_out)
-        raise VOSError(err_msg)
+    if result is None and n_timed_out > 0:
+        err_msg = ('None of the available catalogs returned valid results.'
+                   ' ({0} URL(s) timed out.)'.format(n_timed_out))
+        warnings.warn(err_msg, NoResultsWarning)
 
     return result
 
@@ -291,11 +284,11 @@ class AsyncSearchAll(AsyncBase):
     :ref:`Simple Cone Search Examples <vo-sec-scs-examples>`.
     If no ``timeout`` keyword given, it waits until completion:
 
-    >>> async_allresults = async_search.get(timeout=30)
+    >>> async_allresults = async_search.get(timeout=60)
     >>> all_catalogs = list(async_allresults)
-    >>> first_cone_arr = async_allresults[all_catalogs[0]].array.data
-    >>> first_cone_arr.size
-    36184
+    >>> first_cone_arr = async_allresults[all_catalogs[0]]
+    >>> len(first_cone_arr)
+    74271
 
     """
     def __init__(self, *args, **kwargs):
@@ -437,6 +430,8 @@ def predict_search(url, *args, **kwargs):
     if len(args) != 2:  # pragma: no cover
         raise ConeSearchError('conesearch must have exactly 2 arguments')
 
+    kwargs['service_url'] = url
+    kwargs['return_astropy_table'] = False
     plot = kwargs.pop('plot', False)
     center, radius = args
     sr = _validate_sr(radius)
@@ -445,10 +440,7 @@ def predict_search(url, *args, **kwargs):
 
     # Not using default ConeSearch instance because the attributes are
     # tweaked to match user inputs to this function.
-    cs = ConeSearchClass()
-    cs.URL = url
-    cs.PEDANTIC = kwargs.pop('pedantic', conf.pedantic)
-    cs_pred = RunTimePredictor(cs.query_region, center, **kwargs)
+    cs_pred = RunTimePredictor(ConeSearch.query_region, center, **kwargs)
 
     # Search properties for timer extrapolation
     num_datapoints = 10  # Number of desired data points for extrapolation
@@ -468,10 +460,10 @@ def predict_search(url, *args, **kwargs):
         warnings.warn(
             'Estimated runtime ({0} s) is non-physical with slope of '
             '{1}'.format(t_est, t_coeffs[1]), AstropyUserWarning)
-    elif t_est > cs.TIMEOUT:  # pragma: no cover
+    elif t_est > conf.timeout:  # pragma: no cover
         warnings.warn(
             'Estimated runtime is longer than timeout of '
-            '{0} s'.format(cs.TIMEOUT), AstropyUserWarning)
+            '{0} s'.format(conf.timeout), AstropyUserWarning)
 
     # Predict number of objects
     sr_arr = sorted(cs_pred.results)  # Orig with floating point error