+++ astroquery/vo_conesearch/validator/validate.py
@@ -1,9 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Validate VO Services."""
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-import six
-from six.moves import map
 
 # STDLIB
 import multiprocessing
@@ -88,7 +84,7 @@ def check_conesearch_sites(destdir=os.curdir, verbose=True, parallel=True,
     if url_list == 'default':
         url_list = conf.conesearch_urls
 
-    if (not isinstance(destdir, six.string_types) or len(destdir) == 0 or
+    if (not isinstance(destdir, str) or len(destdir) == 0 or
             os.path.exists(destdir) and not os.path.isdir(destdir)):
         raise IOError('Invalid destination directory')  # pragma: no cover